Improve windows support
In order to make the nvim config more easily maintainable, *nix and
windows platforms needs to be handled with the same config. This is
achieved via guards that detect if nvim is being executed within the
windows platform::

    if has('win32') || has('win64')
	...
    endif

There is a complication with this, however, and that is handling
different conventions for filesystem paths. For example, *nix uses
foreslashes as the path separate while windows uses the backslash. More
testing is needed with this changeset on the windows platform to ensure
that everything works as intended.