Use relative numbers in normal mode
In order to efficiently use vim movement commands, it is necessary to
know the relative position a specific line is. To achieve this, this
changeset sets the ``relativenumber`` option whenever in normal mode.
When in insert mode, absolute line numbers are shown instead.

Also included is a function and keybinding that toggles use of relative
numbers in normal mode.

Lastly, for the changes to work, the ``number`` option must always be
set which this changeset enables directly in ``init.vim``.