Add commented line for new LanguageClient-neovim plugin
The new LanguageClient-neovim moves a lot of functionality into a rust
program. However, as it currently does not work as expected (eg
autocomplete), the old client is being used. As there is a change in how
the new version of the plugin is installed, the install command for the
plugin is left in as a comment so that it is easy to check from time to
time whether the new plugin is good enough for use.