FIX: g:rustfmt_command contains incorrect command
This issue occurred due to switching away manually downloading,
compiling, and installing the rustfmt command via cargo (ie `cargo
install rustfmt-nightly`). Since will now use the precompiled rustfmt
component available via rustup, checking for the existence of the
rustfmt executable on the path, at least on windows, is no longer
correct.

This changeset now only checks for the existence of the rustup
executable. Note that this can break if the rustfmt component is not
included in a particular toolchain release.