New plugin: neomake
This changeset enables Neomake and ALE to be installed at the same time.
However, only ALE is running automatically leaving Neomake to only run
when manually invoked.

Having both plugins installed at the same time is done since Neomake is
able to use some tools that ALE is not able to (eg clippy for rust,
msbuild for c#).