Move vim filetype vimscript to after/plugin
With this change, there is no need to maintain the vim autogroup in
init.vim as the vim.vim after/ftplugin will be loaded automatically.

In general, moving all custom filetype vimscript files to an
after/ftplugin will making maintenance of init.vim much easier due to
not having to manually load a filetype vimscript.