fix(vim-rooter): FindRootDirectory() returns home directory
This is only an issue when using the coc.nvim plugin.

The cause of this issue is the default value of g:rooter_patterns which
has the value of the first element be '.vim/'. Since the home directory
may contain this directory, it is found first.

This change fixes the issue by only having the following array set to
g:rooter_patterns:

['.git/', '.projections.json']