fix: error msg generated w/ missing oceanic-next colorscheme
In cases where the oceanic-next colorscheme is not available (eg
bootstrapping the configuration), or when truecolor is not available (eg
running nvim on a linux virtual console), nvim will generate error
messages advising unable to process the configuration.

This change only loads the oceanic-next colorscheme or the custom colors
plugin if nvim runs where the value of the $TERM env variable is not
'linux' ('linux' is the value of $TERM for the linux virtual console).