Add new plugin: onedark
onedark is a colorscheme for vim/neovim.

Even though the onedark colorscheme is included in the vim-colorschemes
plugin, vim-colorschemes is missing the autoload script needed for the
colorscheme to work properly.

By adding the onedark colorscheme as a separate plugin, the autoload
script then exists ensuring the colorscheme works as expected.