Rename init/ -> init.d/
Since the configuration entrypoint script is named init.vim, it is often
a source of error when referring to either the init.vim file or to a
script inside the init/ directory.

This change renames the init directory to init.d ensuring that it is
clear what is being used.