vim-plug: make vim-plug window open in a horizontal split
It is often hard to read the commit log messages of updated plugins due
to the small width of the vim-plug vertical split window. As such, it
becomes easy to simply ignore the messages. However, ignoring the commit
log messages of updated plugins is dangerous as it becomes easier to be
ignorant of changes that breaks the configuration.

This change configures vim-plug such that the window opens in a
horizontal split making it easier to read the commit messages of updated
plugins.