Add python package list file
In order for some plugins to work (such as ALE), some commands need to
be available. Many of these commands are installable via the python
package manager (ie pip), however there isn't a standard way to specify
exactly which python packages need to be installed.

This change adds a file that lists all packages needed to be installed
in the neovim python virtualenv. This file will be depended on by future
commits which will add the capability to install/upgrade the packages.