Add new global variable: g:python3_bindir
Some python packages installed into the neovim python virtualenv provide
executables. However, these executables are not found on the $PATH and
so cannot be detected normally by some plugins. In order to add these
executables to the $PATH, the path to the neovim python virtualenv bin
directory needs to be defined.

This change adds a new global variable `g:python3_bindir` that holds the
path to the neovim python bin directory.