Add shell scripts to manage neovim python packages
Managing the packages listed in `config/py3pkg.txt` is very much a
manual process; it requires activating the neovim python virtualenv and
then manually upgrading packages via the `pip` command. As such, it is
easy to make a mistake or to forget to upgrade a needed package.

This change adds 2 shell scripts that manage the python packages
listed in `config/py3pkg.txt`. Specifically, the scripts allow to
install, upgrade, or list outdated packages.

Please note that these scripts expect to be passed a single argument
which is the path to the python binary to use in order to
install/upgrade/list outdated packages.