Add custom commands to invoke scripts in bin/
In order to run the scripts that manage the packages installed in the
neovim python virtualenv, the path to the virtualenv's python binary
needs to be passed in as an argument. Doing this by hand is error prone.

This change adds custom commands that invoke the scripts by passing in
the correct path to the neovim python virtualenv's python binary.