Add poetry project files
Managing the python dependencies is a very manual process. As a result,
it is easy to make a mistake or forget where the virtualenv even is
making it more work to maintain.

This change adds poetry project files. This is the first step in
changing the nvim python virtualenv so that it is managed by the poetry
dependency management tool.