Use poetry to get path to virtualenv
Managing the python dependencies is a very manual process. As a result,
it is easy to make a mistake or forget where the virtualenv even is
making it more work to maintain. Also, defining the location of the
virtualenv has to be hard coded into the configuration which presents
challenges when porting to different systems and platforms (since the
location of the virtualenv may need to be in a different location).

This change completes the transition to using the poetry tool to manage
the nvim python virtualenv by using poetry to define the location of the
virtualenv. To accomplish this, the directory given by poetry is cached
and used in future invocations of nvim in order to ensure a fast
startup.