fix: virtualenv not found when cachefile doesn't exist
When the cache file containing the path to the nvim python virtualenv
doesn't exist, the virtualenv is created. However, checkhealth shows
that no virtualenv exists.

The cause of this issue is due to the variable holding the location of
the virtualenv being dropped from memory after the virtualenv is
created.

This change fixes the issue by not dropping the variable containing the
location of the virtualenv from memory.