Add keybinding to fuzzy search current buffer
When searching for text in the current buffer, the text must be inputted
precisely and exactly as written in the file. However, when searching
for text that may or may not be spelled correctly, it becomes more
difficult to find.

This change adds keybindings to invoke Denite to fuzzy search the
current buffer. This makes finding text via fuzzy search more convenient
to use and will make it easier to find text that is spelled slightly
incorrectly.