languageclient: set path to settings json file
In order to enable settings in some language servers, a settings file
needs to be configured. However, a settings file does not exist nor does
the variable g:LanguageClient_settingsPath point to a file that exists.

This change adds a languageclient settings file and configures the
g:LanguageClient_settingsPath variable accordingly. As a result,
language servers are able to be configured from nvim.