vim-plug: increase size of split window
When calling any of the vim-plug commands that opens a new window, the
split takes up half of the window space. However, when there are a lot
of updates, this means that a lot of scrolling up and down must be done
in order to read through the changes of the plugins.

This change makes the new split window take 35 rows in length which is
most of the screen. This makes it easier to read more lines of the
changes implemented by updated plugins.