python: disable using mypy as an ale linter
Some external modules do not have mypy stubs for it. As a result,
importing such libraries will cause mypy to return an error stating that
it cannot find the module even if the module is properly installed.

This change removes mypy from the list of linter tools that ale uses.