Remove plugin: vim-grepper
This plugin allows one to search multiple files for a pattern. However,
The ctrlsf plugin allows one to do this as well but using neovim and
vim8's async support. Thus, there are now 2 plugins installed that both
do the same thing: search for text in multiple files.

This change removes the vim-grepper plugin as the ctrlsf plugin has a
better search experience and better features.