Setup custom node.js environment
The current configuration requires a node.js provider to be properly
setup. However, as the configuration depends on a system node.js being
installed and the neovim javascript package being installed as a global
module -- both requiring manual setup before running neovim -- the
configuration will fail on systems without node.js or if the neovim
javascript package cannot be installed as a global module.

This change checks if a custom node environment exists for the neovim
configuration, and if it does not exist, will download and setup a
custom node environment. This also handles installing the neovim
javascript package.

It should be noted that this change adds the custom node environment
first in the $PATH environment variable. This means that the custom
node.js environment will always be the one used for javascript within
neovim regardless of any other node.js environment that may exist.