languageclient: disable virtualtext for diagnostics
If an error or warning is detected on a line or block of code, the
error/warning message is displayed as virtualtext. However, this is of
limited utility as error messages are usually long enough that they
cannot entirely fit in the space allotted for virtualtext. Also,
sometimes when the error/warning has been fixed, the virtualtext is not
removed.

This change disables using virtualtext to display diagnostic messages.
Diagnostic messages can still be accessed and read either in the
quickfix list or by calling LanguageClient#explainErrorAtPoint().