languageclient: open cursor diagnostics in preview window
Error or warning messages for a line or block of code are added to the
quickfix list. However, if there are many such messages, and the message
for a particular line/block of code is in the middle of the list of
messages, then it can be hard and/or more work than it should be to find
the message.

This change adds a keybinding that opens a preview window containing the
full error or warning message for a particular line or block of code.
This makes it much easier to find out what error or warning has
occurred.