denite: remove command options that no longer exist
Recent versions of the denite plugin included changes that removed some
options from denite commands. This was done to separate the plugin into
two processes: one for the user interface, and another for the plugin
itself. However, in doing so, the plugin removed use of the prompt.nvim,
with the result that some options to denite commands either are
meaningless or cannot be implemented; these options thus no longer
exist, but are included in the denite configuration which results in
error messages complaining that an option is unknown.

This change removes the denite command options that no longer exist
ensuring that the plugin works without errors.

Please note that the change to the denite plugin also removed default
keybindings in search results, but these were not added in this commit.
Keybindings to perform certain actions (eg opening a selected file) will
be added in a future commit.