denite: add keybindings for search result actions
Recent versions of the denite plugin included changes that removed some
options from denite commands. This was done to separate the plugin into
two processes: one for the user interface, and another for the plugin
itself. However, in doing so, the plugin removed action keybindings in
the search results; this is due to the denite plugin no longer using the
prompt.nvim plugin. This means when a denite buffer is opened with
results, no actions besides scrolling up and down can be done.

This change adds action keybindings for denite search result buffers.
Specifically, this allows opening a selected file, selecting multiple
files, filtering results further based on a fuzzy search, and closing
the denite buffer.