denite: use floating window for results
The default view for displaying denite results is in a split. However,
this can be a little disruptive visually as the split opens obscuring
part of the buffer being viewed.

This change uses a floating window to display denite results. With this
change, a floating window appears on screen instead of a split opening.