Denite buffers toggle relativenumber
When entering insert mode in a denite buffer, the relativenumber option
is automatically enabled. This is unnecessary as denite buffers are used
more like popup menus than full fledged buffers meant to be edited.

This change disables enabling or disabling the relativenumber option
when in a buffer with any of the following file types:

* denite

* denite-filter