fix: languageclient no longer working
Even though no there have been no changes to the configuration beyond
updating plugins, the languageclient plugin no longer works. While logs
show that the languageclient does start, it ends abruptly due to an
error.

After research, it was found that expected values of the
``LanguageClient_virtualText`` setting have changed; before the variable
expected either '0' (off) or '1' (on); now the variable expects 'Yes' or
'No'. Changing the value of the variable to either 'Yes' or 'No' causes
the LanguageClient plugin to work again.

This change updates the value of the ``LanguageClient_virtualText``
setting from '0' to 'No'.