plugin: add commented out new variable g:LanguageClient_loggingFile
When troubleshooting the LanguageClient plugin, it is practically
impossible to figure out what is going on without any kind of log.

This change adds and sets a new variable
``g:LanguageClient_loggingFile`` (which is commented out by default) to
hold the value of a path to a log file for the LanguageClient plugin.
When debugging in the future, this line should simply be uncommented
along with the line setting the log level in order to generate a log
file.