New plugin: vim-clap
The fuzzy matcher plugin fruzzy seems unmaintained and as such is a
potential liability wrt issues and/or security. However, if the fruzzy
plugin is removed, then the denite plugin must either use the much
slower default fuzzy matcher, or one of the alternatives listed in the
denite help doc.

This change removes the fruzzy plugin and adds the vim-clap plugin as a
replacement. Included in this change is configuration to make denite use
vim-clap's fuzzy matcher.

Please note: while vim-clap provides roughly the same functionality as
denite, denite is still used to provide fuzzy search menus.