plugin: update source of ale plugin
The source location that's hosting the source of the ale vim plugin has
changed. While, as of this commit, the old location redirects to the new
location, the redirect is not guaranteed to continue in perpetuity.

This change updates the source location of the ale vim plugin to the
current location.