plugin: remove tagbar
The tagbar plugin is increasingly completely or partially lacking
support for an increasing number of filetypes. In addition, the tagbar
plugin does not accomodate either currently or in the forseeable future
modern tools (such as language server clients). As such, the tagbar
plugin is increasingly becoming less useful as time moves on.

As the new vista plugin does accomdate modern tools as well as providing
a way to easily add support for new filetypes, the change removes the
tagbar plugin and updates affected keybinds to use the vista plugin
instead.