fix: non-editor buffers display number gutter
When a non-editor buffer, such as a floating window or a terminal, is
opened, the buffer has the number and relativenumber options set. This
results in the number gutter being displayed which clutters and/or takes
up valuable space in these buffers (especially floating window buffers
which often has limited horizontal space). It is often expected that
these non-editor buffers do not have the number and relativenumber
options set.

This change ensures that non-editor buffers always have the nonumber and
norelativenumber options set. This means that these types of buffers
will never show the number gutter.