Add new global variable: runparts_plugsrc_cmd
Referring to the plugin initialization directory is partially being
manually hardcoded when loading plugins. While this hasn't been a
problem historically due to being isolated to a single region in the
`init.vim` file, implementing other features needing access to this
directory will also need to partially hardcode the location. This may
increase maintenance burden in the future due to errors such as typos.

This change adds a new global variable `runparts_plugsrc_cmd` and
updates the plugin loading code in `init.vim` to refer to this new
variable. Future features than need to do anything with the plugin
initialization directory now only need refer to the new variable instead
of hardcoding the location.