Move complex plugin settings into dedicated files
The configuration of some plugins is either many lines long, or complex,
or both long and complex. This makes browsing through the configuration
file a bit tedious.

This change moves the plugin configuration settings that are
particularly long and/or complex into separate files dedicated to each
plugin.

For example, if the configuration for plugin `example_plug` was moved as
described, then the settings would be moved out of
`init.d/plugin/config.vim` and into a file only for `example_plug` ie
`init.d/plugin/config.d/example_plug.vim` in this example.