Cleanup plugin loading
The `load.vim` file which defines and loads plugins is a bit messy; it
is hard to tell which plugins are pure vimscript, which plugins require
external programs, and which plugins use nvim's remote plugin api.

This change cleans up the `load.vim` file and uses comments to identify
the different types of plugins.