Replace languageclient & ncm2 w/ vim-lsp and asyncomplete
While LanguageClient-neovim (providing a language server client) and
ncm2 (providing auto completion) work reasonably well, there is still
room for some improvement. LanguageClient-neovim has some quirks when
displaying documentation in a floating window (eg unable to focus the
floating window in order to scroll through content), while ncm2 is
implemented primarily in python; it is desired to have as few nvim
plugins as possible utilizing a separate programming language vm
accessed via nvim's remote plugin api.

This change replaces LanguageClient-neovim with vim-lsp which displays
allows focusing preview content in a floating window, and ncm2 with
asyncomplete which is implemented mostly in vimscript.

Please note that these changes needed to be done together due to
requiring additional setup between auto completion and the language
server client. If, for example, vim-lsp was added without also adding
asyncomplete, additional configuration is needed for ncm2 to support
vim-lsp; it this additional configuration was not done, then auto
completion would be broken, and it does not make sense to do the
configuration only for it to be changed anyway in the very next commit.