Use vim-clap for buffer/line/tag fuzzy search
The denite plugin is being used to provide fuzzy search for open
buffers, lines, and vim tags. However, while functional, the behaviour
of normal and insert modes for the fuzzy search buffer is a bit clunky.

Specifically, a denite fuzzy search buffer, when opened, starts off in
either insert or normal mode depending on configuration.  However, the
fuzzy search buffer's specific insert mode also has its own normal mode.
This means once the fuzzy search buffer is in insert mode, in order to
go to the buffer's normal mode, the insert mode first needs to go to
normal mode, and then this normal mode exited out of to get to the
buffer's normal mode. This means once the fuzzy search buffer is in
insert mode, it takes quitting 3 times in order to close the buffer. In
contrast, vim-clap has a keybinding specifically for closing the search
buffer.

This change replaces use of denite with vim-clap for fuzzy search of the
following: open buffers, lines in the current buffer, and vim tags. As
described, this gives the ability to close the search buffer with a
single keybinding.