fix: asyncomplete cfg using multiple autocmd!
The configuration for the asyncomplete plugin defines an autogroup
`lsp_asyncomplete`. The autocommands in this group are all defined as
autocmd!. However, all other autogroups throughout the entire
configuration define only a single autocmd! to begin the group such as:

    augroup example
	autocmd!
	autcmd event1 ...
	autcmd event2 ...
	autcmd event3 ...
    augroup END

This is inconsistent and adds a small increase in cognitive load when
trying to reason about the configuration as a whole.

This change updates the lsp_asyncomplete autogroup so that autocommands
are defined in the same way as described above ie starting with a single
autocmd!.