Add example cfg files based on past config
Previous configurations had different plugins setup and working but for
various reasons had those plugins removed and/or replaced. However, the
configuration for some of these plugins are a bit complex making it hard
to simply remove it. This leads to the configuration containing multiple
commented out sections for plugins that are no longer used at all. This
introduces clutter and makes it harder to find specific plugin settings.

This change adds a new directory `example/` that contains previous
configuration for LanguageClient-neovim, deoplete, deoplete-jedi, jedi,
fzf, ncm2, and opening the terminal buffer in insert mode; where these
settings existed in the configuration commented out, they have been
removed. This makes it easier to add these working configurations back
in if needed in the future. Also, with this change there is now a place
where commented out sections in the configuration can be stored without
cluttering up the config.